% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyply_activate.R
\name{active<-}
\alias{active<-}
\title{Define which of the data-frames in a poly_frame is the `currently-active`
data-frame}
\usage{
active(x) <- value
}
\arguments{
\item{x}{A poly_frame.}

\item{value}{The index of an entry in the poly_frame. This
will become the active data-frame. [Currently: data-frames can be indexed
by integers only].}
}
\description{
Define which of the data-frames in a poly_frame is the `currently-active`
data-frame
}
