% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyply_activate.R
\name{active.poly_frame}
\alias{active.poly_frame}
\title{Return the index of the currently-active data-frame in a poly_frame}
\usage{
\method{active}{poly_frame}(x)
}
\arguments{
\item{x}{A poly_frame.}
}
\description{
Return the index of the currently-active data-frame in a poly_frame
}
