% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyply_activate.R
\name{active}
\alias{active}
\title{Return the index of the currently-active data-frame in the poly_frame}
\usage{
active(x)
}
\arguments{
\item{x}{An object which has a 'active' attribute.}
}
\description{
Return the index of the currently-active data-frame in the poly_frame
}
