% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyply_data.R
\docType{data}
\name{animals}
\alias{animals}
\title{Brain / Body sizes of several mammalian and dinosaur species}
\format{28 * 3 data.frame; columns are 'common_name', 'body', 'brain'.}
\source{
the MASS package dataset `Animals` (but, Dipliodocus ->
  Diplodocus typo was fixed).
}
\usage{
animals
}
\description{
Body weight is in _kg_. Brain weight is in _g_.
}
\keyword{datasets}
