% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyply_poly_frame.R
\name{as_poly_frame.list}
\alias{as_poly_frame.list}
\title{Create a poly_frame object from a list of data-frames}
\usage{
\method{as_poly_frame}{list}(x, merge_fn)
}
\arguments{
\item{x}{A non-empty list of data-frames}

\item{merge_fn}{A function that defines how the data-frames are
to be merged together.}
}
\description{
Create a poly_frame object from a list of data-frames
}
