% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyply_poly_frame.R
\name{as_poly_frame}
\alias{as_poly_frame}
\title{Create a poly_frame object from a list of data-frames}
\usage{
as_poly_frame(x, ...)
}
\description{
Create a poly_frame object from a list of data-frames
}
