% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyply_dplyr_call.R
\name{build_method_from_dplyr_verb}
\alias{build_method_from_dplyr_verb}
\title{Elevates a dplyr-verb from a method on a single data-frame, to a method on
  a poly_frame}
\usage{
build_method_from_dplyr_verb(dplyr_fn)
}
\arguments{
\item{dplyr_fn}{Any dplyr verb that acts on a single data-frame.}
}
\description{
When applied to a poly_frame, the returned function will modify the
  `currently-active data-frame` using the dplyr verb. All other data-frames
  are unmodified and a poly-frame of the same structure is returned (modulo
  any changes to the active data-frame).
}
