% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyply_activate.R
\name{extract_active_df}
\alias{extract_active_df}
\title{Return the contents of the currently-active data-frame from the poly_frame}
\usage{
extract_active_df(x)
}
\arguments{
\item{x}{A poly_frame.}
}
\description{
Return the contents of the currently-active data-frame from the poly_frame
}
