% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyply_filter.R
\name{filter.poly_frame}
\alias{filter.poly_frame}
\title{Filter the active data-frame within a poly_frame}
\usage{
\method{filter}{poly_frame}(.data, ...)
}
\arguments{
\item{.data}{A poly_frame.}

\item{...}{Filtering predicates, as would be used in
dplyr::filter.}
}
\description{
Filter the active data-frame within a poly_frame
}
