% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyply_dplyr_verbs.R
\name{group_by.poly_frame}
\alias{group_by.poly_frame}
\title{Group the entries in the active data-frame by a set of columns, within a
  poly_frame}
\usage{
\method{group_by}{poly_frame}(.data, ...)
}
\arguments{
\item{.data}{A poly_frame.}

\item{...}{Grouping predicates, as would be used in
dplyr::group_by, and other args to be passed to dplyr::group_by.}
}
\description{
Group the entries in the active data-frame by a set of columns, within a
  poly_frame
}
