% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyply_purrr_verbs.R
\name{map_pf}
\alias{map_pf}
\title{Map a function over a (named) list and return a `poly_frame` object}
\usage{
map_pf(.x, .f, ...)
}
\arguments{
\item{.x}{A list or atomic vector. It should have named
elements otherwise a poly_frame cannot be constructed.}

\item{.f}{A function, formula or atomic vector (as in
input to purrr). The function should return a data-frame when applied to
an input data-frame.}

\item{...}{Further arguments to the function `.f`}
}
\description{
See `purrr` for details of `map` functions.
}
