% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyply_dplyr_verbs.R
\name{mutate.poly_frame}
\alias{mutate.poly_frame}
\title{Mutate the active data-frame within a poly_frame}
\usage{
\method{mutate}{poly_frame}(.data, ...)
}
\arguments{
\item{.data}{A poly_frame.}

\item{...}{Mutation predicates, as would be used in
dplyr::mutate.}
}
\description{
Mutate the active data-frame within a poly_frame
}
