% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyply_poly_frame.R
\name{poly_frame}
\alias{poly_frame}
\title{Create a poly_frame object}
\usage{
poly_frame(..., merge_fn)
}
\arguments{
\item{...}{A set of data-frames.}

\item{merge_fn}{A function that defines how the data-frames are
to be merged together.}
}
\description{
Create a poly_frame object
}
