% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyply_activate.R, R/polyply_dplyr_verbs.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{activate}
\alias{reexports}
\alias{filter}
\alias{reexports}
\alias{group_by}
\alias{reexports}
\alias{mutate}
\alias{reexports}
\alias{rename}
\alias{reexports}
\alias{select}
\alias{reexports}
\alias{ungroup}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{filter}}, \code{\link[dplyr]{group_by}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{rename}}, \code{\link[dplyr]{select}}, \code{\link[dplyr]{ungroup}}}

  \item{tidygraph}{\code{\link[tidygraph]{activate}}}
}}

