% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyply_dplyr_verbs.R
\name{rename.poly_frame}
\alias{rename.poly_frame}
\title{Rename columns in the active data-frame within a poly_frame}
\usage{
\method{rename}{poly_frame}(.data, ...)
}
\arguments{
\item{.data}{A poly_frame.}

\item{...}{Renaming details, as would be used in
dplyr::rename.}
}
\description{
Rename columns in the active data-frame within a poly_frame
}
