% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyply_dplyr_verbs.R
\name{select.poly_frame}
\alias{select.poly_frame}
\title{Select within the active data-frame of a poly_frame}
\usage{
\method{select}{poly_frame}(.data, ...)
}
\arguments{
\item{.data}{A poly_frame.}

\item{...}{Columns for selection, as would be used in
dplyr::select.}
}
\description{
Select within the active data-frame of a poly_frame
}
