% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyply_data.R
\docType{data}
\name{taxonomy}
\alias{taxonomy}
\title{Mappings from species-name to taxonomic 'family' and 'order' for a subset of
  the animals present in the `animals` dataset}
\format{23 * 3 data.frame; columns 'species', 'family', 'order'}
\source{
Used taxize::classification over the NCBI database and the
  species names present in `common_to_species`. Code is given in
  `origins_of_the_datasets.Rmd`.
}
\usage{
taxonomy
}
\description{
Note that only mammalian species returned any family / order annotations, so
  you'll have to live without dinosaurs for now.
}
\keyword{datasets}
