% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyply_dplyr_verbs.R
\name{ungroup.poly_frame}
\alias{ungroup.poly_frame}
\title{Ungroup the columns of the active data-frame, within a poly_frame}
\usage{
\method{ungroup}{poly_frame}(x, ...)
}
\arguments{
\item{x}{A poly_frame.}

\item{...}{See dplyr::ungroup}
}
\description{
Ungroup the columns of the active data-frame, within a poly_frame
}
