% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyply_activate.R
\name{update_active_df}
\alias{update_active_df}
\title{Change the contents of the currently-active data-frame in the poly_frame}
\usage{
update_active_df(x, .df)
}
\arguments{
\item{x}{A poly_frame}

\item{.df}{A data-frame to be stored in the active-slot of
the poly_frame \code{x}}
}
\description{
This may alter your ability to merge the data-frames correctly
}
