% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyply_activate.R
\name{activate.poly_frame}
\alias{activate.poly_frame}
\title{Activate a data-frame within a poly_frame}
\usage{
\method{activate}{poly_frame}(.data, what)
}
\arguments{
\item{.data}{A poly_frame.}

\item{what}{An index for that data-frame within the
poly_frame which is to be activated (i.e. become the data-frame that is
prepared for filtering / selecting). This may be specified by an integer
index or by variable-name (i.e. \code{activate(pf, some_df_name)}).}
}
\description{
Activate a data-frame within a poly_frame
}
