% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyply_activate.R
\name{activate.poly_frame}
\alias{activate.poly_frame}
\title{Activate a data-frame within a poly_frame}
\usage{
\method{activate}{poly_frame}(.data, what)
}
\arguments{
\item{.data}{A poly_frame.}

\item{what}{An index for that data-frame within the
poly_frame which is to be activated (ie, become the data-frame that is
prepped for filtering / selecting).}
}
\description{
Activate a data-frame within a poly_frame
}
