% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyply_data.R
\docType{data}
\name{animals}
\alias{animals}
\title{Brain / Body sizes of several mammalian and dinosaur species}
\format{Blah blah blah}
\source{
the MASS package
}
\usage{
animals
}
\description{
Brain / Body sizes of several mammalian and dinosaur species
}
\keyword{datasets}
