% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyply_activate.R, R/polyply_filter.R,
%   R/polyply_mutate.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{activate}
\alias{reexports}
\alias{filter}
\alias{reexports}
\alias{mutate}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{filter}}, \code{\link[dplyr]{mutate}}}

  \item{tidygraph}{\code{\link[tidygraph]{activate}}}
}}

